+++ source/Nuke.Common/Tools/MSpec/MSpecSettings.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Linq;
+
+namespace Nuke.Common.Tools.MSpec
+{
+    public partial class MSpecSettings
+    {
+        private string GetToolPath()
+        {
+            return MSpecTasks.GetToolPath();
+        }
+    }
+}